Serve selected API versions, namespaces
Define what versions should be accessible. Resources are automatically fetched
and routes created.

Default version has prefix '/' (or whatever you choose) and other versions
are hidden in namespace /v1/, /v2/ and so on. Default version has routes
in namespace too, so that an API client might stick with it in the future.