Remove Compat, implement parameter inheriting from resource, specify db_name
Parameters, that are common to more than one action can be defined
in a resource and included by action. This greatly reduces the amount
of written code.

Parameters have new attribute db_name, which can be used when you
want the parameter to have different name in API and in database.

There are two methods that convert parameter names to and fro
parameter and db name: Action#to_db_names and Action#to_param_names.

The first argument to these methods is a hash of parameters to convert.
The second argument is which parameters should be used for conversion,
which can be :input or :output.