Transaction confirmations: fork edit to edit_before and edit_after
edit_before confirms the change of attributes that have been already
altered in the database. Therefore it does nothing if the transaction
succeeds and reverts the original value if it fails.

edit_after is an alias to the original edit and works exactly
the other way around - changes attributes on success, nothing
if failed.