Fix user context switching
Administrator can now (again) switch context and be logged
as another user. A new token for the target user is requested,
the old admin one is remembered. When switching back, the new
token is revoked and the remembered one restored.

Note that while logged as another user, the original token
is not automatically renewed. Therefore, it may expire
in the meantime and the switch back will result in your
being logged off. Also, the target user last activity is not changed,
and neither is your original account's last activity.