Local snapshot mount
Makes possible to mount dataset snapshot to directory in a VPS.
Supports only local mount - the snapshot and the VPS must be
on the same node.

To be able to mount a snapshot, it must be cloned into a dataset.
The temporary dataset is placed in pool's working dataset.
Snapshot reference count was moved from snapshot in branch
to snapshot in pool, because the snapshot cannot be deleted before
it is umounted (and the cloned dataset removed).