VPS subdatasets: do not mount directly
VPS subdatasets are now treated as normal datasets. They are not
mounted directly to /vz/root/<veid>. Instead, normal mount will have
to be created.

The reason for this is to unite behavior of datasets and mainly
because the dataset cannot be mounted directly. The .zfs subdirectory
must not be accessible to the user. All datasets will therefore have
a subdirectory called 'private' and this is the directory that
will be mounted (bind or nfs) to VPSes.