Update the VPS creation wizard
The wizard now uses GET method, except the final page. This way it
is possible to link to any stage in the wizard.

The user first selects an environment, then he's given a list
of locations that have nodes in this environment. The last step is
to choose OS template, hostname and cluster resources. The resources
are preset to the available maximum.

Wizard forms are moved to forms/vps.forms.php.