VPS creation: do not limit cluster resources for admins
When an admin creates a VPS, it is a little ambiguous in which
environment he'll place it, which user will own it and so on. Therefore
admins can now set resources from the minimal allowed value to the
maximum. It does not take their current amount of free resources in
account.