client_params_to_api: send booleans as false
Boolean values are always sent, either true or false. The reason is that
the API may have some default values and if the parameter is not sent at
all, the default value from the API may be used.

This happened when a checkbox representing a parameter was not checked
and the default in the API was true.