Mounter: remote control is not compulsory
When the Mounter is used in a standalone mode, usually from mount action
scripts, the vpsAdmind may not be running or may have disabled remote
control. In these cases, the mount script should not fail, VPS must be
started even if the vpsAdmind is not available at the moment.

This means that mount state will not be accurate without running daemon
and delayed mounts will not work.