Remove user mount scripts
The scripts could be used only when the VPS was running, so only when
the user chose to mount/umount himself.

It cannot be used when the VPS is started, because it is not possible to
nest vzctl commands: vzctl start calls mount action scripts, action
script then has to call the user script, but any vzctl command will fail
to acquire lock, as it is held by the vzctl start command.