Prevent rollback of a dataset with mounted primary snapshots
This temporarily fixes issue #63. Rollback of a dataset is forbidden
if any of its snapshots on a primary/hypervisor pool are mounted to any
VPS.

During rollback, the snapshot will be removed from the primary pool, so to
allow the rollback to proceed, it'd have to change mount types, clone
the snapshots if necessary and remount them.