Vps::Reinstall, Restore: continue if VPS start fails
vpsAdmin cannot guarantee that the VPS template from which it is created
is correct. Nor can it guarantee that the backup itself isn't broken. In
both cases, failure during VPS start meant that the chain had to end up
in fatal state, as they are both destructive and cannot be rollbacked.

Now the VPS won't be able to start, but the user can choose another VPS
template or a different backup to restore from.