Vps::{AddIp,DelIp}: fix resource reallocation
The problem is that ActiveRecords collection .count method behaves
differently than Array.count. AR's count simply returns number of all
records for the query and does not accept a block. The code used the
method as if it behaved like Array.count, which takes the block and
returns the count of items that the block returns true for.