Firewall::Accounting: fix race condition when saving IP traffic
When a new IP address is being registered in iptables, vpsAdmind adds
several rules. If the iptables counter are read at the same time, some
rules for the new IP address may be missing and thus the reading is
incomplete. Some values are not found and it tried to save NULLs to the
database, which essentially deadlocked the firewall code.

The fix is to save zeroes instead of NULLs and register IP addresses in
an atomic step, so that the counters aren't read when a new IP address
is being registered.