Free unused login names
When a user is moved to state hard_delete, all his data on servers is
deleted and the account cannot be revived.However,  his login was still in
the database. Since the login column has a unique key on it, no one else
could ever take a login of a deleted user.

Now, the login is moved to a non-unique column, which enables new users
to use logins of users in state hard_delete.