Web console: limit refresh rate to 20 per second
Before this commit, the console used to send one request after another
to check for updates. When a key was pressed, it was immediately sent
in another request. Now, it sends requests in regular intervals, at max
rate of 20 per second (every 50 ms). Entered keys are buffered and sent
with the next request.

Closes issue #120.