Registration and documentation of mail templates
All mail templates now must be registered using `MailTemplate.register`
method. Use of an unregistered mail template will end with exception.

Mail templates are now identified by an id, a symbol. The id is used
only to identify the template in the code. Template's names can be
parametried and it is these resolved parametrized names that are stored
in the database, just as before this commit. The list of name parameters
is documented and the caller passes the parameters when sending mail.
Resulting template name is automatically resolved.

The reference documentation contains a static file with a list of
registered mail templates, their name, parameters, description and
variables.