Choice fields: use raw value, not its index
In HaveAPI, choice fields are specified using the inclusion validator.
The choices can either be an array of accepted values, or a dictionary,
where keys are sent to the API and values are shown to the user.

Up to now, when an array of accepted values was used, the web UI worked
with indexes of those values. This complicated matters, because you had
to find the actual value to send to the API and then find the index of a
value to show it in a form. Now, it works with the accepted values
directly and no translation is needed.