Reset per-thread current user info when no user is authenticated
When a user is authenticated, his session is stored in
UserSession.current and User.current. These values are accessible only
from the same thread they were stored in. However, since one thread is
used to handle requests repeatedly and a different user can be
authenticated in each request, these values need to be reset on each
request.

This caused a problem only for unauthenticated requests, which did not
overwrite UserSession.current and User.current. The user could not
bypass authorization in the controller, but e.g. transaction chains
created by unauthenticated users were assigned to the user that
was authenticated in the same thread before them.