Store affected VPSes persistently
This allows users to see how many outages their VPS had and what VPSes
are affected by an outage.

The list of affected VPSes is constructed when the outage is announced.
If affected entities are changed after the outage has been announced,
you need to call action Outage.RebuildAffectedVps manually in order to
rebuild the list.