Differentiate OpenVZ and vpsAdminOS templates
When creating a new VPS as an unprivileged user, available templates are
listed based on selected location and available hypervisor types. When
reinstalling an existing VPS, only templates for appropriate hypervisor
type are shown.

When we'll be dealing with migrations of OpenVZ and vpsAdminOS VPSes, we
could add additional column to table `os_templates` to select target
template for vpsAdminOS hypervisors.