Split nodectld into libnodectld and nodectld, nodectl script
libnodectld contains all the code, nodectld then contains the executable
and CLI. The reason for splitting the code is because we need to be able
to use nodectld standalone mode for scripting and osctl hooks. However,
it is impossible to `require 'nodectld/standalone'`, because the path is
not known, it is somewhere in Nix store. So, nodectl has a new
subcommand `script`, which can run any Ruby script with libnodectld in
path. This is done by making libnodectld a dependency of nodectl.

Scripts can then use shebang like:

    #!/run/current-system/sw/bin/nodectl script