nodectld: use /run/nodectl/nodectl script as shebang in osctl hooks
Instead of using /run/current-system/sw/bin/nodectl, osctl hooks are now
using symlink /run/nodectl/nodectl as shebang. The reason for this changed
is to simplify development of these hook scripts, when you wish to use
development version of nodectl, instead of the one that was bundled with
the OS. Now, all you have to do is to point /run/nodectl/nodectl to the
development version and all hook scripts will use it immediately.