libnodectld: fix reenabling of device access
nodectld always attempts to delete all devices that are not supposed to
be allowed and add devices that should be allowed. However, osctl exits
with an error when you try to add a device that's already there. So we
check the return code and error message to handle this case.