api: fix IP resource reallocation
If a VPS was created without v4 or v6 address, the respective resource
wasn't allocated at all. That's all right, except when the VPS was
deleted, resources for *both* v4 and v6 were reallocated in
`Vps::DelIp`. In `Vps::DelIp`, we expected all resources to be already
allocated and that they're just being changed. Instead, the resource
allocation for the missing IP version was created using `edit()`
transaction confirmation, which left the allocation in an unconfirmed
state `confirm_create`. When the user created another VPS with
the previously missing IP version, `Vps::Create` mistakenly though it
has just created the allocation and used `create()` transaction
confirmation on it. This meant that the reallocation changes weren't
applied. The resource was assigned, but was not accounted. That's
because `create()` just confirms the row, it does not apply changes.