api: Vps::Clone::VzToOs: use OsTemplate.distribution and OsTemplate.version
When cloning from OpenVZ to vpsAdminOS, vpsAdmin has to find an
appropriate OS template. Since not all OpenVZ templates have
counterparts, we need to have a way of configuring mapping from OpenVZ
templates to vpsAdminOS ones.

OpenVZ templates are identified by attribute `name`, while vpsAdminOS
templates use attributes `distribution`, `version` and so on. So, the
mapping can be configured by setting `distribution` and `version`
attributes of OpenVZ templates, which are otherwise not used in any way.
For example, for OpenVZ template with name=debian-lenny, you could set
distribution=debian, version=8.0, as that is the oldest debian template
available on vpsAdminOS.