api: Vps::Clone::VzToOs: fix clone of VPS with subdatasets
Cloning of VPS with subdatasets resulted in the subdatasets being
unreachably mounted. It happened as follows: first, all targets datasets
are created, mounted, and private/ subdirectories are created. Then all
datasets are transfered via zfs send/zfs recv -F. The -F means that
current data in the dataset is discarded, which in case of the top-level
dataset includes all subdataset mountpoints. The datasets then remained
mounted, but their mountpoints were unreachable.

When cloning VPS, we don't actually need to create the private/
subdirectories, because they will be in the received zfs stream. The
datasets then don't have to be mounted at all before the VPS is to be
started.