Route network addresses via configured host addresses
Routing via on-interface host addresses is necessary when there's not a
host address from the routed network on the interface.

Note that this routing mechanism works only on vpsAdminOS and
integration with VPS migration/swap/clone is not complete. It's not an
issue at the moment since these operations do not work between
vpsAdminOS nodes at all.