api: introduce cluster resource packages
Cluster resources are no longer to be manipulated independently
per-user. Instead, administrators will define sets of resource packages
which will then be assigned to users. Every package carries a set of
cluster resources which the user gets when the package is assigned to
him. In addition, every user has his own personal resource package, so
that individual configuration is still possible.