api: Vps::Migrate::Base: fix removal of old IP addresses
When replacing IP addresses during a migration to a different location,
the old addresses have to be removed on the destination node. This is
because VPS configs are copied from the source node, including the IP
addresses that have to be removed. Removing them on the source node
at this point was pointless.