api: when clearning netif, removed routed-via addresses first
We delete host addresses first, so when some route is being routed via a
host addresses that is removed, osctl will simply remove such routes,
but vpsAdmin doesn't know about it and tries to remove it itself, which
fails. So now we delete routed-via addresses first and the rest later.
It still doesn't properly handle order of removed addresses, i.e. it may
be possible to break this code as well, but it should cover more cases
than before.