libnodectld,vpsadmind: pipe socat to zfs recv
Using exec in socat does not work well, because when the socket is
closed, it either kills zfs recv, or when end-close is used, it closes
fds and exists. However, we need to wait for zfs recv to finish before
continuing, and socat does not appear to support this, so we connect
them using pipes.