libnodectld: do not remove mbuffer log file on recv
The code did not work anyways, as the mbuffer | zfs recv command is
forked and the nodectld transaction exits. Therefore it attempted to
delete the log file even before it was created.

We could extend blocking_fork to accept a code block which would be
invoked after the subprocess finishes. For now we'll keep the log files
to see how it performs.