nixos: database: disable the query cache
It is buggy and making vpsAdmin unreliable. We've been getting strange
errors on prepared statements, which apparently got cached in a weird
way. E.g. in nodectld:

    ArgumentError: invalid value for BigDecimal(): "2021-11-01T16:34:36"

This happened on varchar fields which should not be parsed as
BigDecimal. Only some random rows were affected by this and resetting
the cache helped. We therefore disable it completely.