libnodectld: protect group snapshots against unexpected crashes
It sometimes happens that there is a failure while communicating with
the database, which can end up in an exception or downright segfault of
the ruby interpreter. This is a problem when we're making snapshots,
because the snapshots get created, but are not saved. After nodectld is
restarted, it makes new snapshots and saves those. So the snapshots
created before the crash remain on disk, but are not tracked and must be
manually deleted.

We now save state locally after the snapshots are made, so if nodectld
crashes, it can restore from this state and not create new snapshots.

(cherry picked from commit ebeeb2a0ba2c88e3e378ee96d577994b01a979e8)