libnodectld: populate config overwrites previous config
It could happen that a VPS migration fails and leaves VPS config on the
target node. When the migration is re-run, populate config is called
again and since the config from the previous is still there and
popupated, it fails to add the network interfaces and causes the
migration to fail. Preexisting configs are now ignored and overwritten.

(cherry picked from commit ac66706ebb0bd800df02c13b6b0d5de3caf4f3b5)