api, libnodectld: on-demand user namespace map use/disuse
User namespace maps are on nodes represented by osctl users. Each VPS
has exactly one user namespace map and the osctl user has to exist
before the VPS can be created. Until now, the existence of osctl users
has been controlled by the API, which stored this information using
UserNamespaceMapPool model. If a user was needed, it was created. When
it was no longer needed, it was destroyed. However, this solution
required that locks be held on the user namespace map, so that we could
ensure atomicity of the operation. This meant that only one migration
per user could be running.

The new solution moves the control from the API to individual nodes. The
API will always fire transactions that either declare a user is needed
or that he's no longer needed. It's now up to nodectld to track if a user
has to be created or destroyed.

(cherry picked from commit 71b86e363d36403c1bb37bd58c8f8f1292f2af04)