api, libnodectld: move user namespace maps to VPS
Having user namespace maps on dataset in pools was incorrect. It is the
VPS which is using the mapping and the fact that we also set zfs
uidmap/gidmap is an implementation detail. In the future, we might use
kernel's ID-mapped mounts instead of our ZFS patch. Since mapping is set
only on VPS datasets, the attribute was completely removed from dataset
in pools. Changing user namespace map on a VPS will automatically
reconfigure all its datasets.