api: handle dataset expansion reduction
If the refquota is changed by admin or user, we can resolve existing
expansion only if the new refquota is actually equal to or lesser than
the original quota. It is possible that the quota was lowered, but is
still above the original quota. In that case, we record the returned
space and keep the expansion alive.