libnodectld: insert OOM report tasks outside of transaction
VPS can have hundreds or thousands of tasks and inserting them all in a
transaction can cause the transaction to fail and be retried. Reasons
for this are unclear. Since there is a finite number of transaction
attempts, it occasionaly led to nodectld crashing. Inserting the tasks
outside of the transaction should avoid the crash.