api, console_router, libnodectld: use quorum queues for non-exclusive queues
While we do not need these queues to be durable and can afford to lose
messages, this strategy does not work in a cluster. When a node with
classic queues goes down, the queues are lost without clients knowing
about it. The auto-recovery mechanism in client works only when the
client itself gets disconnected, which might not be the case.