api: resolve dataset expansion when new refquota exceeds referenced space
It can happen that the user has a sufficient amount of additional space
in his account, but the dataset was expanded past it. Close the dataset
expansion when the user sets the refquota higher than it was before
the expansion, even if the new refquota is lower than the expansion
made it. We check only that the new refquota is greater than the currently
referenced space.