api, webui: replace dataset expansion deadline with over refquota time tracking
Deadlines as a static date time was insufficient and flawed for
long-term expansions. Imagine that a dataset was expanded as it was
nearly full, but it didn't exceed its original refquota and stayed that
way for more than the deadline allowed. Such scenario is perfectly
acceptable, but if the dataset should later indeed exceed its original
refquota, the deadline would be in effect and the VPS could be stopped.

Tracking the time in which datasets are over the original refquota
handles this case better, giving users more time to solve the issue.