api: check if the user can login before issuing OAuth2 authorization code
We check if the user's account is not locked out only before the actual
login, after authentication was successful. For OAuth2, that is when the
authorization code has already been issued, so the error will happen in
the token endpoint, which isn't designed to report user errors.

We therefore check if the user can log in before authorization code is
issued and that way we can report the error while the user is still on
the authorization page.