console_router, libnodectld: reset response on each request
Since `@response` was reset only when RpcClient instance was created, long
lived clients mistakenly returned responses to previous requests when a
timeout occurred.

After `@condition.wait` returns, the client evaluates whether it got a
reply or not. Since `@response` was set to the last request's response,
the client didn't recognize a timeout and returned invalid response.