api: destroy auth tokens when totp is skipped, refactor types
`Operations::Authentication::Password` creates AuthToken if the user has
TOTP enabled and when the caller decides to skip TOTP, the AuthToken was
left behind, which resulted in it expiring and UserFailedLogin record
was created.

Since to destroy AuthToken we need the record itself, make sure that
AuthResult#auth_token is always AuthToken and not String.