api, libnodectld, webui: secondary internal DNS zones
Until now, internal zones had to be primary, i.e. when DNS records were
changed, all servers were updated. This patch allows for internal zones
to be secondaries. Changing records will now only update primary
instances and the secondary instances will perform zone transfers.

This is a preliminary step for supporting DNSSEC on internal zones,
because for the bind automation to work, the keys can be only on one
primary server and the rest have to be secondaries.