libnodectld: fix named.conf syntax for TSIG in allow-transfer
We cannot combine IPs with keys as in primaries {}. For allow-transfer,
it is either the TSIG key or the IP address, so we opt for the TSIG key
if it is set. It'd be best to explore this in the future, so that both
the TSIG key and the IP address would have to match.