api, libnodectld: increment/decrement confirmation for multiple attributes
One increment/decrement transaction confirmation can now handle multiple
attributes of the same record and with custom increment/decrement value.
The previous implementation allowed incrementing/decrementing only a
single attribute and by a constant of 1.

This feature is for now disabled within the API, as it requires all
nodectld instances to be updated before we can use it.