api: configurable per-IP version dynamic DNS URLs
Clients with both IPv4 and IPv6 connectivity need to use update domains
that have only A or AAAA records set, not both. For example, if we'd use
a domain that has both A and AAAA records, then either only IPv4 or IPv6
could be updated, depending on what IP version would the client use by
default.

While with some programs can specify the IP version to be used, e.g.
curl has this ability, not all of them might and having separate domains
will be more reliable.