webui: inline DNS record add form
This should be more understandable to users. The standalone form was
rather confusing, because what we need most of the time is record name,
type and content. It was unintuitive that the less used fields as TTL
and priority were mixed in. Even though it was in zone file order, I
found it to be more complicated.

The downside of this approach is that now the user can't set a comment
when creating a new record. Comments can still be set in the record edit
page, which retains the original form.