Version 4.0.0
While vpsAdmin is no longer versioned and we simply use a git revision,
vpsadmin-client is a Ruby gem that is published on rubygems.org. As
such, it needs to conform to semantic versioning. Until now, we've been
making releases using prerelease/beta mechanism and we've added current
timestamp as the version. This is problematic for gems depending on the
client, as rubygems won't consider prereleases for auto-updates.

To resolve this, we will simply conform to semantic versioning again.
All of vpsAdmin component versions will be updated, although in reality,
only the client will be using it. The current version was bumped to
4.0.0 to signify this change.