api, libnodectld, nixos: support for OS template configs
OS template configs can be used to preconfigure subdatasets, mounts and
features, which will be applied when new VPS is created. The intended
use-case for now is NixOS impermanence templates. The following config
will ensure its functionality:

	datasets:
	- name: "/"
	  properties:
	    refquota: 75%
	- name: nix
	  properties:
	    refquota: 25%
	mounts:
	- mountpoint: /persistent
	  dataset: "/"
	- mountpoint: /nix
	  dataset: nix
	features:
	  impermanence: true

Only a single level of subdatasets is supported. refquota can specified
in percent of allocated diskspace, or in number of MiB.

While the container image with NixOS impermanence already contains all
of this configuration, vpsAdmin wouldn't be aware of it. nodectld will
clear any mounts from the template and vpsAdmin will add its own.