api, libnodectld: handle VPS exit events and disable autostart
When the VPS is halted from the inside, autostart in osctl remained
enabled. When osctld was restarted, e.g. during an update, it could
restart the VPS.

nodectld listens for ct_exit events in osctld and forwards them to the
supervisor. The supervisor logs all halt/reboot events and in case of
halt, next status update will disable autostart if the VPS is still
stopped.