api: remember user's last next multi factor authentication choice
When using TOTP, users can choose when will the next TOTP code be
required on the same device. The default is to require the code on each
login, but users can opt to trust a device for up to a month a for that
time, vpsAdmin does not ask for a TOTP code. Next time vpsAdmin needs to
ask for the TOTP code, it now remembers the user's previous choice for
how long should the device be trusted.