webui: ensure the entry page is within history
Without this, the history always contains one less page. When we
traverse back, the latest page used to be lost.webui: ensure the entry
page is within history

Without this, the history always contains one less page. When we
traverse back, the latest page used to be lost.webui: ensure the entry
page is within history

Without this, the history always contains one less page. When we
traverse back, the latest page used to be lost.webui: ensure the entry
page is within history

Without this, the history always contains one less page. When we
traverse back, the latest page used to be lost.