api, libnodectld, webui: update template using osctl ct set distribution
`osctl ct set image-config` was used, because container images can
contain specific osctl config fragments. Changing the configuration
required the VPS to be stopped. In practice, there are no config
fragments that would be needed. In the past, the config contained
arguments to /sbin/init, which was necessary to run systemd on systems
with cgroups v1. However, since osctld automatically adjusts this
systemd argument based on what cgroups version is available, the config
no longer needs to be applied.