libnodectld: read os-release using osctl ct cat
Reading os-release from chroot does not take subdatasets and their
mounts into account. This is an issue for NixOS with impermanence,
where /nix is stored in subdataset. Since /etc/os-release is a symlink
to the Nix store, it is not possible to read it from chroot.

While slower, osctl ct cat reads files from the container's own mount
namespace.