api, webui: configurable multi-factor authentication
Especially with passkeys, it makes sense to give users a way to
configure passkeys, but keep multi-factor authentication disabled until
desired, rather then enable it automatically when passkey exists. This
is because passkeys might be tied to a specific system and would prevent
the user from setting up passkeys on other systems.