api: fix race when incrementing DNS zone serial
It appears that `dns_zone.increment!(:serial)` sends
`serial = serial + 1` to the database, but does not work with the new value
from the database and instead increments the attribute locally. When two
requests are updating the zone at the same time, this can lead to a race
condition where the serial is correctly incremented twice in the
database, but the locally incremented attribute will end up the same for
both requests.

`DnsZone#increment_serial` now reads the new serial value from the
database.