api, libnodectld-v5, nodectl-v5: refactor network accounting into VPS statuses
Network interface stats are now obtained using libvirt instead of
reading files in /sys. Libvirt domains are already accessed when
checking VPS statuses, so it makes sense to add network accounting it.

Note that network accounting used to be send in batches, each batch
contained stats for N network interfaces from multiple VPS. The current
solution will lead to more SQL queries, since now the
monitors/accounting will be updated for each VPS individually. It can
still be optimized in the future.

In addition, network accounting is updated right away instead of caching
the values for a configured amount of time within nodectld. This will
lead to more SQL queries as well and might need optimization.

As it is, the net_monitor and net_accounting routing keys are now unused
by nodectld-v5.