api, libnodectld-v5, nixos: support for rescue mode of libvirt VPS
There is one action to enter rescue mode and another action to leave the
rescue mode. Unlike with osctld, libvirt VPS will remain in rescue mode
until it is explicitly disabled.

A temporary volume is created for the rescue system. This volume is
stored in vpsAdmin. vpsAdmin actions touching the VPS must take the
rescue volume in mind.