api, libnodectld-v5, libnodectld: track precise VPS state
We used to track only if the VPS is running or not, but the VPS can
be in many other states, mainly paused (for libvirt) and frozen (for
vpsAdminOS).

libvirt also tracks a reason for how the VPS got into the state, but we
do not process it for now.