libnodectld: retry RPC read requests
Sometimes, the client does not receive a response from the supervisor
within the hard timeout period. The request/response is either lost or
the supervisor fails; the reason is unclear. Since all of the RPC
client's requests are actually for obtaining information from the
database, nodectld does not have to crash, it can resend the request.

For this purpose, the hard timeout was shortened to 60s.